rename: rename multi module
멀티 모듈 이름 수정
