+++ module-consumer/build.gradle
@@ -2,8 +2,10 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.kafka:spring-kafka'
-    implementation project(':module-domain:module-rdbms')
+    implementation project(':module-core:module-jpa')
 
     runtimeOnly 'com.h2database:h2'
     runtimeOnly 'com.mysql:mysql-connector-j'
+
+    testImplementation project(':module-core:module-jpa')
 }
