+++ module-consumer/src/main/java/com/example/moduleconsumer/service/MemberService.java
@@ -1,9 +1,9 @@
 package com.example.moduleconsumer.service;
 
 import com.example.moduleconsumer.request.CreateMemberRequest;
-import com.example.modulerdbms.domain.Member;
-import com.example.modulerdbms.entity.MemberEntity;
-import com.example.modulerdbms.repository.MemberRepository;
+import com.example.modulejpa.domain.Member;
+import com.example.modulejpa.entity.MemberEntity;
+import com.example.modulejpa.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
