+++ module-core/module-jpa/build.gradle
