+++ module-core/module-jpa/src/main/java/com/example/modulejpa/domain/Coupon.java
@@ -1,22 +1,16 @@
-package com.example.modulerdbms.entity;
+package com.example.modulejpa.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
-
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-@Entity
-@Table(name = "coupon")
-public class CouponEntity {
+public class Coupon {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private Long memberId;
 }
