+++ module-core/module-jpa/src/main/java/com/example/modulejpa/entity/MemberEntity.java
@@ -1,16 +1,22 @@
-package com.example.modulerdbms.domain;
+package com.example.modulejpa.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import javax.persistence.*;
+
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
+@Entity
 @Builder
-public class Member {
+@Table(name = "member")
+public class MemberEntity {
 
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
     private String email;
