+++ module-core/module-jpa/src/main/java/com/example/modulejpa/repository/MemberJpaRepository.java
@@ -1,6 +1,6 @@
-package com.example.modulerdbms.repository;
+package com.example.modulejpa.repository;
 
-import com.example.modulerdbms.entity.MemberEntity;
+import com.example.modulejpa.entity.MemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
