+++ module-core/module-redis/.gitignore
