+++ module-core/module-redis/build.gradle
