+++ module-core/module-redis/src/main/java/com/example/moduleredis/repository/AppliedUserRepository.java
