+++ module-producer/src/test/java/com/example/moduleproducer/service/ApplyServiceTest.java
@@ -3,7 +3,6 @@ package com.example.moduleproducer.service;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.data.redis.core.RedisTemplate;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
@@ -14,29 +13,6 @@ class ApplyServiceTest {
 
     @Autowired
     private ApplyService applyService;
-    @Autowired
-    private RedisTemplate<String, String> redisTemplate;
-
-    @Test
-    public void manyRequest() throws Exception {
-        int threadCount = 1000;
-
-        ExecutorService executorService = Executors.newFixedThreadPool(32);
-        CountDownLatch countDownLatch = new CountDownLatch(threadCount);
-
-        for (int i = 0; i < threadCount; i++) {
-            long memberId = i;
-            executorService.submit(() -> {
-                try {
-                    applyService.apply(memberId);
-                } finally {
-                    countDownLatch.countDown();
-                }
-            });
-        }
-
-        countDownLatch.await();
-    }
 
     @Test
     public void manyRequestByUniqueUser() throws Exception {
@@ -56,13 +32,5 @@ class ApplyServiceTest {
         }
 
         countDownLatch.await();
-    }
-
-    @Test
-    public void redisIncrement() {
-        Long coupon_count = redisTemplate.opsForValue().increment("coupon_count");
-        String coupon_count1 = redisTemplate.opsForValue().get("coupon_count");
-        System.out.println(coupon_count);
-        System.out.println(coupon_count1);
     }
 }
