+++ settings.gradle
@@ -2,8 +2,6 @@ rootProject.name = 'kafka-practice'
 include 'module-core'
 include 'module-consumer'
 include 'module-producer'
-include 'module-domain:module-rdbms'
-findProject(':module-domain:module-rdbms')?.name = 'module-rdbms'
-include 'module-domain:module-redis'
-findProject(':module-domain:module-redis')?.name = 'module-redis'
+include 'module-core:module-jpa'
+include 'module-core:module-redis'
